Finish chat refresh reconciliation cleanup
